factored out merging
